+++ mayan/apps/credentials/model_mixins.py
@@ -1,15 +1,7 @@
-from mayan.apps.events.decorators import method_event
-from mayan.apps.events.event_managers import EventManagerMethodAfter
-
 from .events import event_credential_used
 
 
 class StoredCredentialBusinessLogicMixin:
-    @method_event(
-        event_manager_class=EventManagerMethodAfter,
-        event=event_credential_used,
-        target='self'
-    )
     def get_backend_data(self):
         obj = super().get_backend_data()
 
@@ -18,5 +10,11 @@ class StoredCredentialBusinessLogicMixin:
 
             if hasattr(backend_class, 'post_processing'):
                 obj = backend_class.post_processing(obj=obj)
+
+        if not getattr(self, '_event_ignore_credential_used', False):
+            actor = getattr(self, '_event_actor', None)
+
+            if actor or self.pk:
+                event_credential_used.commit(actor=actor, target=self)
 
         return obj