+++ mayan/apps/source_emails/tests/test_source_backends.py
@@ -1,3 +1,4 @@
+from mayan.apps.credentials.events import event_credential_used
 from mayan.apps.documents.events import (
     event_document_created, event_document_file_created,
     event_document_file_edited, event_document_version_created,
@@ -459,39 +460,59 @@ class IMAPSourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_false(self):
         test_document_count = Document.objects.count()
@@ -521,39 +542,59 @@ class IMAPSourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_none(self):
         test_document_count = Document.objects.count()
@@ -583,39 +624,59 @@ class IMAPSourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_true(self):
         test_document_count = Document.objects.count()
@@ -645,39 +706,54 @@ class IMAPSourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 8)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, self._test_document_type)
+        self.assertEqual(events[2].actor, test_document)
+        self.assertEqual(events[2].target, test_document)
+        self.assertEqual(events[2].verb, event_document_created.id)
 
         self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].actor, test_document_file)
+        self.assertEqual(events[3].target, test_document_file)
+        self.assertEqual(events[3].verb, event_document_file_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_version)
+        self.assertEqual(events[5].target, test_document_version)
+        self.assertEqual(events[5].verb, event_document_version_created.id)
+
+        self.assertEqual(events[6].action_object, test_document_version)
+        self.assertEqual(events[6].actor, test_document_version_page)
+        self.assertEqual(events[6].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[6].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[7].action_object, None)
+        self.assertEqual(events[7].actor, self._test_stored_credential)
+        self.assertEqual(events[7].target, self._test_stored_credential)
+        self.assertEqual(events[7].verb, event_credential_used.id)
 
 
 class POP3SourceBackendActionDocumentUploadTestCase(
@@ -712,39 +788,59 @@ class POP3SourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_false(self):
         test_document_count = Document.objects.count()
@@ -774,39 +870,59 @@ class POP3SourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_none(self):
         test_document_count = Document.objects.count()
@@ -836,39 +952,59 @@ class POP3SourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 9)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, None)
+        self.assertEqual(events[2].actor, self._test_stored_credential)
+        self.assertEqual(events[2].target, self._test_stored_credential)
+        self.assertEqual(events[2].verb, event_credential_used.id)
 
-        self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].action_object, self._test_document_type)
+        self.assertEqual(events[3].actor, test_document)
+        self.assertEqual(events[3].target, test_document)
+        self.assertEqual(events[3].verb, event_document_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_created.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_file)
+        self.assertEqual(events[5].target, test_document_file)
+        self.assertEqual(events[5].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[6].action_object, test_document)
+        self.assertEqual(events[6].actor, test_document_version)
+        self.assertEqual(events[6].target, test_document_version)
+        self.assertEqual(events[6].verb, event_document_version_created.id)
+
+        self.assertEqual(events[7].action_object, test_document_version)
+        self.assertEqual(events[7].actor, test_document_version_page)
+        self.assertEqual(events[7].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[7].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[8].action_object, None)
+        self.assertEqual(events[8].actor, self._test_stored_credential)
+        self.assertEqual(events[8].target, self._test_stored_credential)
+        self.assertEqual(events[8].verb, event_credential_used.id)
 
     def test_dry_run_true(self):
         test_document_count = Document.objects.count()
@@ -898,36 +1034,51 @@ class POP3SourceBackendActionDocumentUploadTestCase(
         )
 
         events = self._get_test_events()
-        self.assertEqual(events.count(), 5)
+        self.assertEqual(events.count(), 8)
 
         test_document = Document.objects.first()
         test_document_file = test_document.file_latest
         test_document_version = test_document.version_active
         test_document_version_page = test_document_version.pages.first()
 
-        self.assertEqual(events[0].action_object, self._test_document_type)
-        self.assertEqual(events[0].actor, test_document)
-        self.assertEqual(events[0].target, test_document)
-        self.assertEqual(events[0].verb, event_document_created.id)
+        self.assertEqual(events[0].action_object, None)
+        self.assertEqual(events[0].actor, self._test_stored_credential)
+        self.assertEqual(events[0].target, self._test_stored_credential)
+        self.assertEqual(events[0].verb, event_credential_used.id)
 
-        self.assertEqual(events[1].action_object, test_document)
-        self.assertEqual(events[1].actor, test_document_file)
-        self.assertEqual(events[1].target, test_document_file)
-        self.assertEqual(events[1].verb, event_document_file_created.id)
+        self.assertEqual(events[1].action_object, None)
+        self.assertEqual(events[1].actor, self._test_stored_credential)
+        self.assertEqual(events[1].target, self._test_stored_credential)
+        self.assertEqual(events[1].verb, event_credential_used.id)
 
-        self.assertEqual(events[2].action_object, test_document)
-        self.assertEqual(events[2].actor, test_document_file)
-        self.assertEqual(events[2].target, test_document_file)
-        self.assertEqual(events[2].verb, event_document_file_edited.id)
+        self.assertEqual(events[2].action_object, self._test_document_type)
+        self.assertEqual(events[2].actor, test_document)
+        self.assertEqual(events[2].target, test_document)
+        self.assertEqual(events[2].verb, event_document_created.id)
 
         self.assertEqual(events[3].action_object, test_document)
-        self.assertEqual(events[3].actor, test_document_version)
-        self.assertEqual(events[3].target, test_document_version)
-        self.assertEqual(events[3].verb, event_document_version_created.id)
+        self.assertEqual(events[3].actor, test_document_file)
+        self.assertEqual(events[3].target, test_document_file)
+        self.assertEqual(events[3].verb, event_document_file_created.id)
 
-        self.assertEqual(events[4].action_object, test_document_version)
-        self.assertEqual(events[4].actor, test_document_version_page)
-        self.assertEqual(events[4].target, test_document_version_page)
+        self.assertEqual(events[4].action_object, test_document)
+        self.assertEqual(events[4].actor, test_document_file)
+        self.assertEqual(events[4].target, test_document_file)
+        self.assertEqual(events[4].verb, event_document_file_edited.id)
+
+        self.assertEqual(events[5].action_object, test_document)
+        self.assertEqual(events[5].actor, test_document_version)
+        self.assertEqual(events[5].target, test_document_version)
+        self.assertEqual(events[5].verb, event_document_version_created.id)
+
+        self.assertEqual(events[6].action_object, test_document_version)
+        self.assertEqual(events[6].actor, test_document_version_page)
+        self.assertEqual(events[6].target, test_document_version_page)
         self.assertEqual(
-            events[4].verb, event_document_version_page_created.id
+            events[6].verb, event_document_version_page_created.id
         )
+
+        self.assertEqual(events[7].action_object, None)
+        self.assertEqual(events[7].actor, self._test_stored_credential)
+        self.assertEqual(events[7].target, self._test_stored_credential)
+        self.assertEqual(events[7].verb, event_credential_used.id)